+++ test/test_py_air_fetcher.py
@@ -11,7 +11,7 @@ def test_metrics_no_host_provided(caplog):
     """
     Error logs explain that the purifier host has to be provided through an env var
     """
-    assert py_air_fetcher.get_status() is None
+    assert py_air_fetcher.get_reading() is None
     assert "Please specify the host address" in caplog.text
     assert py_air_fetcher.HOST_ENV_VAR in caplog.text
 
@@ -19,14 +19,14 @@ def test_metrics_no_host_provided(caplog):
 def test_metrics_pyairctrl_failure(mock_http_client, caplog):
     """Error logs explain that there was a failure getting the status from pyairctrl"""
     mock_http_client["get_status"].side_effect = Exception("Some foobar error")
-    assert py_air_fetcher.get_status(host="1.2.3.4", protocol="http") is None
+    assert py_air_fetcher.get_reading(host="1.2.3.4", protocol="http") is None
     assert "Could not read values from air control device" in caplog.text
     assert "Some foobar error" in caplog.text
 
 
 def test_metrics_unknown_client(caplog):
     """Error logs explain that the chosen protocol is unknown"""
-    assert py_air_fetcher.get_status(host="1.2.3.4", protocol="foobar") is None
+    assert py_air_fetcher.get_reading(host="1.2.3.4", protocol="foobar") is None
     assert "Unknown protocol 'foobar'" in caplog.text
 
 
@@ -51,16 +51,14 @@ def test_get_client_plain_coap_protocol(mock_plain_coap_client):
 
 
 @pytest.mark.usefixtures("mock_http_client")
-def test_get_status():
-    assert py_air_fetcher.get_status("1.2.3.4", "http") == metrics.AirControlStatus(
+def test_get_reading():
+    assert py_air_fetcher.get_reading("1.2.3.4", "http") == metrics.TargetReading(
         status=metrics.Status(fan_speed=0, iaql=1, is_manual=True, is_on=True, pm25=2),
-        filters=metrics.Filters(
-            filters={
-                "0": metrics.Filter(hours=0, filter_type=""),
-                "1": metrics.Filter(hours=185, filter_type="A3"),
-                "2": metrics.Filter(hours=2228, filter_type="C7"),
-            },
-        ),
+        filters={
+            "0": metrics.Filter(hours=0, filter_type=""),
+            "1": metrics.Filter(hours=185, filter_type="A3"),
+            "2": metrics.Filter(hours=2228, filter_type="C7"),
+        },
     )